+++ src/components/icons/index.ts
@@ -101,6 +101,8 @@ import StateError from 'src/assets/svg/mp_info_lt_state-error.svg?react'
 import StateNoResults from 'src/assets/svg/mp_info_lt_state-no-results.svg?react'
 import Back from 'src/assets/svg/mp_act_lt_back.svg?react'
 import Flag from 'src/assets/svg/flag.svg?react'
+import SubmitFeedback from 'src/assets/svg/mp_act_lt_submit-feedback.svg?react'
+import Boost from 'src/assets/svg/mp_info_lt_boost.svg?react'
 
 export {
   AbSplit,
@@ -190,6 +192,7 @@ export {
   ShieldKeyholeIcon,
   SignoutIcon,
   SplitIcon,
+  SubmitFeedback,
   SystemAlertsIcon,
   TextWidget,
   TransformationsIcon,
@@ -206,4 +209,5 @@ export {
   StateError,
   StateNoResults,
   Back,
+  Boost,
 }